Passing functions as props from Children to Parent
The SearchBar component isn't responsible for making search requests. The parent/caller of the SearchBar is, so the parent should define what happens when a search is performed. This method should be passed as a prop to the search bar component, which in turn should just call it at the appropriate time when a search occurs, and the execution happens through the parent.